Restructure types and definitions
