+++ bokehjs/src/lib/models/tools/gestures/box_select_tool.ts
@@ -1,10 +1,10 @@
 import {RegionSelectTool, RegionSelectToolView} from "./region_select_tool"
-import type {KeyBinding} from "../tool"
 import {BoxAnnotation} from "../../annotations/box_annotation"
 import {Coordinate} from "../../coordinates/coordinate"
 import type {Scale} from "../../scales/scale"
 import type {IconLike} from "../../common/kinds"
 import type {FactorLike} from "../../ranges/factor_range"
+import type {KeyBinding} from "core/keyboard"
 import type * as p from "core/properties"
 import type {SelectionMode, CoordinateUnits} from "core/enums"
 import {Dimensions, BoxOrigin} from "core/enums"
