+++ bokehjs/src/lib/models/tools/gestures/pan_tool.ts
@@ -1,9 +1,9 @@
 import {GestureTool, GestureToolView} from "./gesture_tool"
-import type {KeyBinding} from "../tool"
 import type {RangeInfo, RangeState} from "../../plots/range_manager"
 import {MenuItem} from "../../ui/menus"
 import type {MenuItemLike} from "../../ui/menus"
 import type {IconLike} from "../../common/kinds"
+import type {KeyBinding} from "core/keyboard"
 import type * as p from "core/properties"
 import type {PanEvent} from "core/ui_events"
 import {assert} from "core/util/assert"
