+++ bokehjs/src/lib/models/tools/gestures/region_select_tool.ts
@@ -1,8 +1,8 @@
 import {SelectTool, SelectToolView} from "./select_tool"
-import type {KeyBinding} from "../tool"
 import type {BoxAnnotation} from "../../annotations/box_annotation"
 import type {PolyAnnotation} from "../../annotations/poly_annotation"
 import type {DataRendererView} from "../../renderers/data_renderer"
+import type {KeyBinding} from "core/keyboard"
 import {RegionSelectionMode} from "core/enums"
 import type {SelectionMode} from "core/enums"
 import type {Geometry} from "core/geometry"
